\name{GastricCancer}
\alias{GastricCancer}
\title{Gastric Cancer Data}
\description{
There are 90 observations on 4 variables.
Times is the survival times in years.
Status is the censoring status. 
ax is the covariate that satisfy proportional hazards assumptions. 
zo is the covariate that indicating the two treatments, which we suppose follow the Y-P model.
For more details please see the reference below.

Data are from 
Ying, Z., Jung, SH, and Wei, LJ (1995). 
Median regression analysis with censored data.
Journal of the American Statistical Association,
90, 178-184.
}
\usage{data(GastricCancer)}
\format{A data frame containing 90 observations on 4 variables:
\tabular{rl}{
        [,1] \tab "times"\cr
        [,2] \tab "status"\cr
        [,3] \tab "ax"\cr
        [,4] \tab "zo"}
}
\references{
Ying, Z., Jung, SH, and Wei, LJ
(1995). Median regression analysis with censored data.
Journal of the American Statistical Association, 90, 178-184.
}
\keyword{datasets}
